feat: reorganize and logs
