+++ .gitignore
@@ -2,3 +2,4 @@
 node_modules
 storage.json
 build
+logs.txt
