+++ package.json
@@ -29,6 +29,7 @@
     "@types/express": "^4.17.17",
     "eslint": "^8.18.0",
     "eslint-config-xo": "^0.41.0",
-    "eslint-plugin-promise": "^6.0.0"
+    "eslint-plugin-promise": "^6.0.0",
+    "prettier": "^2.8.7"
   }
 }
